Change parse tests to use golden data
Change the parse unit tests to use a golden test data pattern. Expected
result data can be generated by passing the -update flag to go test,
which triggers an override of the expected result files stored in tdata.
Whenever this is done, any changes should be carefully checked manually.

Once the expected data is committed, and when go test is triggered
without the -update flag, parsed data is compared to the previously
committed expected data.